+++ src/orcaset/__init__.py
@@ -1,7 +1,14 @@
 # Copyright (c) 2026 Orcaset Inc.
 # SPDX-License-Identifier: SSPL-1.0
 
-from orcaset.context import Context, CycleError, DepNode
+from orcaset.context import Context, ConvergenceError, CycleError, DepNode
+from orcaset.date_series import (
+    DateExtendSeries,
+    DateMap2Series,
+    DateMapSeries,
+    DateSeries,
+    DateSeriesBase,
+)
 from orcaset.formatters import (
     DateFormatter,
     ValueFormatter,
@@ -11,22 +18,35 @@ from orcaset.formatters import (
 )
 from orcaset.maybe import Maybe, Na, add_values, combine_values, isna, map2_some, map_some
 from orcaset.period import Period, date_union, period_union
-from orcaset.query import DayCount, accrual, accrual_or, exact, exact_or, last
-from orcaset.rule import Demand, KeyedRule, Rule, Step, get, get_at
+from orcaset.period_series import (
+    PeriodExtendSeries,
+    PeriodMap2Series,
+    PeriodMapSeries,
+    PeriodSeries,
+    PeriodSeriesBase,
+)
+from orcaset.query import DayCount, accrual, accrual_or, covered, exact, exact_or, last
+from orcaset.rule import (
+    Demand,
+    Iterate,
+    KeyedRule,
+    Rule,
+    Step,
+    abs_distance,
+    get,
+    get_at,
+    maybe_abs_distance,
+)
 from orcaset.series import (
     BaseSeries,
     CellFactory,
     CellsFn,
     CellStream,
-    DateSeries,
-    DateSeriesBase,
     Key,
     Map2Series,
     MapItemsSeries,
     MapNSeries,
     MapSeries,
-    PeriodSeries,
-    PeriodSeriesBase,
     QueryFn,
     Replayable,
     Series,
@@ -52,8 +72,12 @@ __all__ = [
     "CellStream",
     "CellsFn",
     "Context",
+    "ConvergenceError",
     "CycleError",
+    "DateExtendSeries",
     "DateFormatter",
+    "DateMap2Series",
+    "DateMapSeries",
     "DateSeries",
     "DateSeriesBase",
     "DateValue",
@@ -62,6 +86,7 @@ __all__ = [
     "DepNode",
     "Group",
     "GroupRow",
+    "Iterate",
     "Key",
     "KeyedRule",
     "LineRow",
@@ -72,6 +97,9 @@ __all__ = [
     "Maybe",
     "Na",
     "Period",
+    "PeriodExtendSeries",
+    "PeriodMap2Series",
+    "PeriodMapSeries",
     "PeriodSeries",
     "PeriodSeriesBase",
     "PeriodValue",
@@ -86,10 +114,12 @@ __all__ = [
     "Total",
     "TotalRow",
     "ValueFormatter",
+    "abs_distance",
     "accrual",
     "accrual_or",
     "add_values",
     "combine_values",
+    "covered",
     "csv_table",
     "date_union",
     "exact",
@@ -102,5 +132,6 @@ __all__ = [
     "map2_some",
     "map_some",
     "markdown_table",
+    "maybe_abs_distance",
     "period_union",
 ]